requirements.txt
flask==3.1.1
prometheus_client
google-cloud-secret-manager
requests
pytest
black
ruff
mypy
coverage